[BUGFIX] Changed from int to long, since some sites uses a lot of bandwidth
Some sites used very, very much bandwith, and the total number of Bytes, could
not be store in a 32bit in value, so the ReducerOutput is not LongWritable

+++ src/CalculateBytesReducer.java
@@ -1,5 +1,6 @@
 import java.io.IOException;
 import java.util.*;
+import java.math.BigInteger;
 
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
@@ -8,14 +9,14 @@ import org.apache.hadoop.mapreduce.Reducer;
 import org.apache.hadoop.mapreduce.Reducer.Context;
 
 
-public class CalculateBytesReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class CalculateBytesReducer extends Reducer<Text, IntWritable, Text, LongWritable> {
 	
 	public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-		int totalBytes = 0;
+		long totalBytes = 0L;
 		for (IntWritable value : values) {
-			totalBytes += value.get();
+			totalBytes += (long)value.get();
 		}
-		context.write(key, new IntWritable(totalBytes));
+		context.write(key, new LongWritable(totalBytes));
 	}
 	
 }
